+++ index.html
@@ -231,6 +231,9 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-<script src="js/main.js"></script>
+{# <script type="module" src="js/generate-photo-list.js"></script> #}
+{# <script type="module" src="js/get-random-int.js"></script> #}
+{# <script type="module" src="js/max-allowed-string-length.js"></script> #}
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
