+++ js/max-allowed-string-length.js
@@ -1,0 +1,4 @@
+function maxAllowedStringLength (string, maxLength) {
+  return string.length <= maxLength;
+}
+export {maxAllowedStringLength};
